yh bro

+++ src/app/api/menu-items/route.js
@@ -1,27 +1,34 @@
  import { Menuitem } from "@/app/models/Menu-items"
 import mongoose from "mongoose"
+import { isAdmin } from "../auth/[...nextauth]/route"
  export  async function POST(req){
-
-    const data = await req.json()
     mongoose.connect(process.env.MONGO_URL)
-     const menu = await Menuitem.create(data)
-     return Response.json(menu)
+    const data = await req.json()
+if(await isAdmin()){
+   const menu = await Menuitem.create(data)
+   return Response.json(menu)
+
+} else{
+
+  return  Response.json({})
+}
 
  }
 
 
 
- export async function PUT(req){
+ export async function PUT(req){ 
     mongoose.connect(process.env.MONGO_URL)
     const {_id, ...data} = await req.json()
       await Menuitem.findByIdAndUpdate(_id ,data)
       return Response.json(true)
  }
 export  async function GET(){
-mongoose.connect(process.env.MONGO_URL)
-   return Response.json(
-       await Menuitem.find()
-   )  
+   mongoose.connect(process.env.MONGO_URL)
+      return Response.json(
+         await Menuitem.find()
+      )  
+      
 
 }
 
@@ -29,8 +36,11 @@ export async function DELETE(req){
    mongoose.connect(process.env.MONGO_URL)
     const url = new URL(req.url);
     const _id = url.searchParams.get("_id")
-    await Menuitem.deleteOne({_id})
-    return Response.json(true)
+    if(await isAdmin()){
+       await Menuitem.deleteOne({_id})
+      }
+      return Response.json(true)
+
    
  }
